Refactor / 코드 리팩토링, modules 폴더 삭제
