+++ src/redux/configStore.js
@@ -1,5 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
-import posts from "../modules/posts";
+import posts from "./posts";
 
 const store = configureStore({
   reducer: { posts: posts.reducer },
